build(jan03): Phase 0 compatibility patches for modern MSVC x86
Bulk patches applied during Silent Storm port Phase 0 to get the 271-cpp
Main DLL + Game.exe + 11 engine subprojects compiling under MSVC 19.50 /x86.

Categories:
  - 196 'if (CDynamicCast<T> v(arg))' → 'CDynamicCast<T> v(arg); if (v)'
    plus 104 'else if' variants — modern C++ rejects this init-in-condition
    pattern when the type has multiple ctors.
  - 30+ 'std::list::insert(end())' / 'push_back()' → emplace_back / emplace
  - 5 'inline float fabs/cos/sin/acos/asin(float)' in Misc/tools.h removed
    — std <cmath> provides identical noexcept overloads.
  - MemoryMngr/DumbPow2Alloc.h rewritten to defer to <new>.
  - Misc/Geom.h SHMatrix: explicit default ctor (anon union of CVec4
    deletes implicit).
  - Misc/Basic2.h: global non-template operator== /!= for CPtr/CObj/CMObj
    to disambiguate 50+ C2666 ambiguities.
  - Misc/EventsBase.h: cache typeid pointer in ctor to allow forward-decl
    TParam in dtor.
  - Misc/tools.h: IsInSet uses manual loop instead of std::find.
  - Main/StdAfx.h pulls all NDb data headers (MSVC eagerly instantiates
    CDBPtr<T>::operator&).
  - Main/Interface.cpp: dead CMouseCaptureHandler (referenced absent
    IWindow/IInterface types) wrapped in '#if 0'.
  - DBFormat/Data*.cpp: missing full-type includes for GetTable<T>() sites.
  - MemoryMngr/Trace.cpp: GetSourceLine call uses &out-pointers.
  - FileIO/BasicChunk1.cpp: emplace_back for chunks list.
  - Input/Input.cpp: emplace(end()) for messages list.

See port repo: docs/upstream-patches.md and docs/patches/*.md for the
detailed per-file list.

+++ Soft/Andy/Jan03/a5dll/Main/iMissionUI.cpp
@@ -80,7 +80,7 @@ inline wstring ConvertLineBreaks( const wstring &szStr )
 				if ( i != szStr.end() && *i == L'\n' )
 					++i;
 				continue;
-			case 133: // symbol L'…'
+			case 133: // symbol L'ï¿½'
 				szRet += L"...";
 				break;
 			default:
@@ -862,8 +862,8 @@ void CMissionUI::UpdateEnemies()
 		}
 
 		const NUI::SPoint &sSize = pIcon->GetSize();
-		vScreenPos.x = max( min( vScreenPos.x, sViewRect.x2 ), sViewRect.x1 );
-		vScreenPos.y = max( min( vScreenPos.y, sViewRect.y2 ), sViewRect.y1 );
+		vScreenPos.x = max( min( vScreenPos.x, (float)sViewRect.x2 ), (float)sViewRect.x1 );  // silent-storm-port: explicit casts for min/max
+		vScreenPos.y = max( min( vScreenPos.y, (float)sViewRect.y2 ), (float)sViewRect.y1 );
 
 		pIcon->Set( pEnemy, bVisible, fAngle );
 		pIcon->SetPosition( SPoint( vScreenPos.x, vScreenPos.y ) );
